[ATfL] A script which generates modulefiles directory tree that could be pointed at with MODULEPATH
This is a proof of concept script which generates a directory tree
with modulefiles. It may be used later when the RPM and DEB
post-install shell scripts are being implemented.

Currently it can be used as such:

In the ATfL tarball (as created by build.sh) a new directory called
'arm' is being created containing the mkmodulesdir.sh script. This
script is being amended by build.sh in order to adjust the ATfL
version info. After unpacking the tarball, an user can execute this
script, it will create three directories in the current directory:

moduledeps
modulefiles
moduleglobals

These contain all the files needed to load ATfL as a module.

Assuming this script has been executed in the atfl/arm directory in
the user's $HOME, ATfL can be loaded as such:

$ MODULEPATH=$HOME/atfl/arm/modulefiles module load atfl/0.0